"Pale Blue Dot"
I've noticed that there are many apps or website features that show the user little tidbits of trivia that vary from subject to subject. And that gave me the idea that, if the user spends time learning bit by bit, they will feel knowledgeable about many different genres of study.

So, to "make the world a better place", I decided to create my own website that updates every day with a fantastic picture of the latest news about the Planet, as well as accompany said picture with a short piece that details what is happening within the picture. For I feel that if more people know about our planet's natural activity, more people will feel more inclined to keep it clean, more people will be amazed to find out that somewhere, at this very moment, something amazing is happening.

If more people find wonder in this world of ours, I believe that it can be on its way to become a better place.